+++ test/unit/dropin.js
@@ -506,20 +506,15 @@ describe('Dropin', function () {
       });
     });
 
-    it('does not load checkout.js if script tag with paypal checkout id is already on the page', function (done) {
+    it('does not load checkout.js if global paypal object is already on the page', function (done) {
       var instance = new Dropin(this.dropinOptions);
-      var script = document.createElement('script');
 
       global.paypal = this.paypalCheckout;
-
-      script.id = 'braintree-dropin-paypal-checkout-script';
-      document.body.appendChild(script);
 
       this.dropinOptions.merchantConfiguration.paypal = {flow: 'vault'};
 
       instance._initialize(function () {
         expect(instance._loadScript).to.not.have.been.called;
-        document.body.removeChild(script);
 
         done();
       });