+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 CHANGELOG
 =========
 
+unreleased
+----------
+- Simplify check for checkout.js on the merchant's page
+
 1.7.0
 ------
 - Add data collector
@@ -27,7 +31,7 @@ CHANGELOG
 - Add `clearSelectedPaymentMethod` to remove selected payment method
 
 1.4.0
-------
+-----
 - Add `paymentOptionSelected` event
 - Add support for PayPal and PayPal credit in the script tag integration
 - Add support for locale and payment option priority in the script tag integration