+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 CHANGELOG
 =========
 
-x.x.x
-------
+unreleased
+----------
+- Simplify check for checkout.js on the merchant's page
 - Allow useraction to be set for PayPal button.
 
 1.7.0
@@ -31,7 +32,7 @@ x.x.x
 - Add `clearSelectedPaymentMethod` to remove selected payment method
 
 1.4.0
-------
+-----
 - Add `paymentOptionSelected` event
 - Add support for PayPal and PayPal credit in the script tag integration
 - Add support for locale and payment option priority in the script tag integration